Moving xmss related functions to a single place
